Code improvements in ResourceHelper - No impact on functionalities - SpellChecker now defines corpusFormat for dataset training of txt files - ResourceHelper now uses ResourceFormat - ResourceHelper now more lightweight, annotator specific manipulation moved to annotator - Config management moved to annotator specific
