+++ src/main/scala/com/jsl/nlp/annotators/ner/regex/NERRegexApproach.scala
@@ -2,6 +2,7 @@ package com.jsl.nlp.annotators.ner.regex
 
 import com.jsl.nlp.AnnotatorApproach
 import com.jsl.nlp.util.io.ResourceHelper
+import com.typesafe.config.{Config, ConfigFactory}
 import org.apache.spark.ml.param.{Param, StringArrayParam}
 import org.apache.spark.ml.util.{DefaultParamsReadable, Identifiable}
 import org.apache.spark.sql.Dataset
@@ -34,13 +35,36 @@ class NERRegexApproach(override val uid: String) extends AnnotatorApproach[NERRe
 }
 
 object NERRegexApproach extends DefaultParamsReadable[NERRegexApproach] {
+
+  private val config: Config = ConfigFactory.load
+
+  /**
+    *
+    * @param entitiesPath The path to load the dictionary from
+    * @param fileFormat The format of the file specified at the path
+    * @return The dictionary as a Map
+    */
+  private def retrieveEntityDict(entitiesPath: String = "__default",
+                         fileFormat: String = config.getString("nlp.entityRecognition.format")
+                        ): Map[String, String] = {
+    val filePath = if (entitiesPath == "__default") config.getString("nlp.entityRecognition.file") else entitiesPath
+    ResourceHelper.parseKeyValueText(filePath, fileFormat, ":")
+  }
+
+  private def retrieveEntityDicts(files: Array[String],
+                          fileFormat: String = config.getString("nlp.entityRecognition.format")
+                         ): Map[String, String] = {
+
+    files.map( f => ResourceHelper.parseKeyValueText(f, fileFormat, ":") ).foldRight(Map[String, String]())( (m1, m2) => m1 ++ m2)
+  }
+
   def train(path: String = "__default"): NERRegexModel = {
-    val entityDictionary = ResourceHelper.retrieveEntityDict(path)
+    val entityDictionary = retrieveEntityDict(path)
     new NERRegexModel().setModel(entityDictionary)
   }
 
   def train(files: Array[String]): NERRegexModel = {
-    val entityDictionary = ResourceHelper.retrieveEntityDicts(files)
+    val entityDictionary = retrieveEntityDicts(files)
     new NERRegexModel().setModel(entityDictionary)
   }
   def train(dict: Map[String, String]): NERRegexModel = {
