+++ src/main/scala/com/jsl/nlp/util/io/ResourceHelper.scala
@@ -2,11 +2,16 @@ package com.jsl.nlp.util.io
 
 import java.io.{File, FileNotFoundException, InputStream}
 
+import com.jsl.nlp.annotators.{Normalizer, RegexTokenizer}
+import com.jsl.nlp.{DocumentAssembler, Finisher}
 import com.jsl.nlp.annotators.common.{TaggedSentence, TaggedWord}
-import com.typesafe.config.{Config, ConfigFactory}
+import com.jsl.nlp.util.io.ResourceFormat._
+import org.apache.spark.ml.Pipeline
+import org.apache.spark.sql.SparkSession
 
 import scala.collection.mutable.{ArrayBuffer, Map => MMap}
 import scala.io.Source
+
 
 /**
   * Created by saif on 28/04/17.
@@ -17,11 +22,10 @@ import scala.io.Source
   */
 object ResourceHelper {
 
-  /** uses config tunning parameters */
-  private val config: Config = ConfigFactory.load
+  private val spark: SparkSession = SparkSession.builder().getOrCreate()
 
   /** Structure for a SourceStream coming from compiled content */
-  private case class SourceStream(resource: String) {
+  case class SourceStream(resource: String) {
     val pipe: Option[InputStream] = try {
       val touchSource = getClass.getResourceAsStream(resource)
       Source.fromInputStream(touchSource)("UTF-8").getLines().take(1)
@@ -35,23 +39,9 @@ object ResourceHelper {
       pipe.foreach(_.close())
     }
   }
-  /**Standard splitter for general purpose sentences*/
-  private def wordTagSplitter(sentence: String, tagSeparator: Char):
-  Array[TaggedWord] = {
-    val taggedWords: ArrayBuffer[TaggedWord] = ArrayBuffer()
-      sentence.split("\\s+").foreach { token => {
-        val tagSplit: Array[String] = token.split('|').filter(_.nonEmpty)
-        if (tagSplit.length == 2) {
-          val word = tagSplit(0)
-          val tag = tagSplit(1)
-          taggedWords.append(TaggedWord(word, tag))
-        }
-      }}
-    taggedWords.toArray
-  }
 
-  /** Checks wether a path points to directory */
-  private def pathIsDirectory(path: String): Boolean = {
+  /** Checks whether a path points to directory */
+  def pathIsDirectory(path: String): Boolean = {
     //ToDo: Improve me???
     if (path.contains(".txt")) false else true
   }
@@ -67,11 +57,11 @@ object ResourceHelper {
     */
   def parseKeyValuesText(
                          source: String,
-                         format: String,
+                         format: Format,
                          keySep: String,
                          valueSep: String): Map[String, Array[String]] = {
     format match {
-      case "txt" =>
+      case TXT =>
         val sourceStream = SourceStream(source)
         val res = sourceStream.content.getLines.map (line => {
           val kv = line.split (keySep).map (_.trim)
@@ -94,11 +84,11 @@ object ResourceHelper {
     */
   def parseKeyValueText(
                           source: String,
-                          format: String,
+                          format: Format,
                           keySep: String
                         ): Map[String, String] = {
     format match {
-      case "txt" =>
+      case TXT =>
         val sourceStream = SourceStream(source)
         val res = sourceStream.content.getLines.map (line => {
           val kv = line.split (keySep).map (_.trim)
@@ -118,10 +108,10 @@ object ResourceHelper {
     */
   def parseLinesText(
                       source: String,
-                      format: String
+                      format: Format
                      ): Array[String] = {
     format match {
-      case "txt" =>
+      case TXT =>
         val sourceStream = SourceStream(source)
         val res = sourceStream.content.getLines.toArray
         sourceStream.close()
@@ -139,11 +129,11 @@ object ResourceHelper {
     */
   def parseTupleText(
                          source: String,
-                         format: String,
+                         format: Format,
                          keySep: String
                        ): Array[(String, String)] = {
     format match {
-      case "txt" =>
+      case TXT =>
         val sourceStream = SourceStream(source)
         val res = sourceStream.content.getLines.map (line => {
           val kv = line.split (keySep).map (_.trim)
@@ -164,11 +154,11 @@ object ResourceHelper {
     */
   def flattenRevertValuesAsKeys(
                                  source: String,
-                                 format: String,
+                                 format: Format,
                                  keySep: String,
                                  valueSep: String): Map[String, String] = {
     format match {
-      case "txt" =>
+      case TXT =>
         val m: MMap[String, String] = MMap()
         val sourceStream = SourceStream(source)
         sourceStream.content.getLines.foreach(line => {
@@ -185,14 +175,14 @@ object ResourceHelper {
 
   def wordCount(
                  source: String,
-                 format: String,
+                 format: Format,
                  m: MMap[String, Int] = MMap.empty[String, Int].withDefaultValue(0),
                  clean: Boolean = true,
                  prefix: Option[String] = None,
                  f: Option[(List[String] => List[String])] = None
                ): MMap[String, Int] = {
     format match {
-      case "txt" =>
+      case TXT =>
         val regex = if (clean) "[a-zA-Z]+".r else "\\S+".r
         if (pathIsDirectory(source)) {
           try {
@@ -234,165 +224,35 @@ object ResourceHelper {
         }
         if (m.isEmpty) throw new FileNotFoundException("Word count dictionary for spell checker does not exist or is empty")
         m
-      case _ => throw new IllegalArgumentException("Only txt supported as a file format")
+      case TXTDS =>
+        import spark.implicits._
+        val dataset = spark.read.textFile(source)
+        val wordCount = spark.sparkContext.broadcast(MMap.empty[String, Int].withDefaultValue(0))
+        val documentAssembler = new DocumentAssembler()
+          .setInputCol("value")
+        val tokenizer = new RegexTokenizer()
+          .setInputCols("document")
+          .setOutputCol("token")
+        val normalizer = new Normalizer()
+          .setInputCols("token")
+          .setOutputCol("normal")
+        val finisher = new Finisher()
+          .setInputCols("normal")
+          .setOutputCols("finished")
+          .setAnnotationSplitSymbol("--")
+        new Pipeline()
+          .setStages(Array(documentAssembler, tokenizer, normalizer, finisher))
+          .fit(dataset)
+          .transform(dataset)
+          .select("finished").as[String]
+          .foreach(text => text.split("--").foreach(t => {
+            wordCount.value(t) += 1
+          }))
+        val result = wordCount.value
+        wordCount.destroy()
+        result
+      case _ => throw new IllegalArgumentException("format not available for word count")
     }
-  }
-
-  /**
-    * Parses CORPUS for tagged sentences
-    * @param text String to process
-    * @param tagSeparator Separator for provided String
-    * @return A list of [[TaggedSentence]]
-    */
-  def parsePOSCorpusFromText(
-                              text: String,
-                              tagSeparator: Char
-                            ): Array[TaggedSentence] = {
-    val sentences: ArrayBuffer[Array[TaggedWord]] = ArrayBuffer()
-    text.split("\n").filter(_.nonEmpty).foreach{sentence =>
-      sentences.append(wordTagSplitter(sentence, tagSeparator))
-    }
-    sentences.map(TaggedSentence(_)).toArray
-  }
-
-  /**
-    * Parses CORPUS for tagged sentence from any compiled source
-    * @param source for compiled corpuses, if any
-    * @param tagSeparator Tag separator for processing
-    * @return
-    */
-  def parsePOSCorpusFromSource(
-                  source: String,
-                  tagSeparator: Char
-                ): Array[TaggedSentence] = {
-    val sourceStream = SourceStream(source)
-    val lines =
-      sourceStream.content.getLines()
-        .filter(_.nonEmpty)
-        .map(sentence => wordTagSplitter(sentence, tagSeparator))
-        .toArray
-    sourceStream.close()
-    lines.map(TaggedSentence(_))
-  }
-
-  /**
-    * Reads POS Corpus from an entire directory of compiled sources
-    * @param dirName compiled content only
-    * @param tagSeparator tag separator for all corpuses
-    * @param fileLimit limit of files to read. Can help clutter, overfitting
-    * @return
-    */
-  def parsePOSCorpusFromDir(
-                             dirName: String,
-                             tagSeparator: Char,
-                             fileLimit: Int
-                           ): Array[TaggedSentence] = {
-    try {
-      new File(dirName).listFiles()
-        .take(fileLimit)
-        .flatMap(fileName => parsePOSCorpusFromSource(fileName.toString, tagSeparator))
-    } catch {
-      case _: NullPointerException =>
-        val sourceStream = SourceStream(dirName)
-        val res = sourceStream
-          .content
-          .getLines()
-          .take(fileLimit)
-          .flatMap(fileName => parsePOSCorpusFromSource(dirName + "/" + fileName, tagSeparator))
-          .toArray
-        sourceStream.close()
-        res
-    }
-  }
-
-  /**
-    * Retrieves Corpuses from configured compiled directory set in configuration
-    * @param fileLimit files limit to read
-    * @return TaggedSentences for POS training
-    */
-  def retrievePOSCorpus(
-                         posDirOrFilePath: String = "__default",
-                         fileLimit: Int = 50
-                       ): Array[TaggedSentence] = {
-    //ToDo support multiple formats in corpus source
-    val dirOrFilePath = if (posDirOrFilePath == "__default") config.getString("nlp.posDict.dir") else posDirOrFilePath
-    val posFormat = config.getString("nlp.posDict.format")
-    val posSeparator = config.getString("nlp.posDict.separator").head
-    val result = {
-      if (pathIsDirectory(dirOrFilePath)) parsePOSCorpusFromDir(dirOrFilePath, posSeparator, fileLimit)
-      else parsePOSCorpusFromSource(dirOrFilePath, posSeparator)
-    }
-    if (result.isEmpty) throw new Exception("Empty corpus for POS")
-    result
-  }
-
-  /**
-    * Retrieves Lemma dictionary from configured compiled source set in configuration
-    * @return a Dictionary for lemmas
-    */
-  def retrieveLemmaDict(
-                         lemmaFilePath: String = "__default",
-                         lemmaFormat: String = config.getString("nlp.lemmaDict.format"),
-                         lemmaKeySep: String = config.getString("nlp.lemmaDict.kvSeparator"),
-                         lemmaValSep: String = config.getString("nlp.lemmaDict.vSeparator")
-                       ): Map[String, String] = {
-    val filePath = if (lemmaFilePath == "__default") config.getString("nlp.lemmaDict.file") else lemmaFilePath
-    ResourceHelper.flattenRevertValuesAsKeys(filePath, lemmaFormat, lemmaKeySep, lemmaValSep)
-  }
-
-  /**
-    * Sentiment dictionaries from compiled sources set in configuration
-    * @return Sentiment dictionary
-    */
-  def retrieveSentimentDict(sentFilePath: String = "__default"): Map[String, String] = {
-    val filePath = if (sentFilePath == "__default") config.getString("nlp.sentimentDict.file") else sentFilePath
-    val sentFormat = config.getString("nlp.sentimentDict.format")
-    val sentSeparator = config.getString("nlp.sentimentDict.separator")
-    ResourceHelper.parseKeyValueText(filePath, sentFormat, sentSeparator)
-  }
-
-  /**
-    * Regex matcher rules
-    * @param rulesFilePath
-    * @param rulesFormat
-    * @param rulesSeparator
-    * @return
-    */
-  def retrieveRegexMatchRules(
-                               rulesFilePath: String = "__default",
-                               rulesFormat: String = config.getString("nlp.regexMatcher.format"),
-                               rulesSeparator: String = config.getString("nlp.regexMatcher.separator")
-                             ): Array[(String, String)] = {
-    val filePath = if (rulesFilePath == "__default") config.getString("nlp.regexMatcher.file") else rulesFilePath
-    ResourceHelper.parseTupleText(filePath, rulesFormat, rulesSeparator)
-  }
-
-  def retrieveEntityExtractorPhrases(
-                                     entitiesPath: String = "__default",
-                                     fileFormat: String = config.getString("nlp.entityExtractor.format")
-                                   ): Array[String] = {
-    val filePath = if (entitiesPath == "__default") config.getString("nlp.entityExtractor.file") else entitiesPath
-    ResourceHelper.parseLinesText(filePath, fileFormat)
-  }
-
-  /**
-    *
-    * @param entitiesPath The path to load the dictionary from
-    * @param fileFormat The format of the file specified at the path
-    * @return The dictionary as a Map
-    */
-  def retrieveEntityDict(entitiesPath: String = "__default",
-                         fileFormat: String = config.getString("nlp.entityRecognition.format")
-                        ): Map[String, String] = {
-     val filePath = if (entitiesPath == "__default") config.getString("nlp.entityRecognition.file") else entitiesPath
-    ResourceHelper.parseKeyValueText(filePath, fileFormat, ":")
-  }
-
-  def retrieveEntityDicts(files: Array[String],
-                          fileFormat: String = config.getString("nlp.entityRecognition.format")
-                         ): Map[String, String] = {
-    
-     files.map( f => ResourceHelper.parseKeyValueText(f, fileFormat, ":") ).foldRight(Map[String, String]())( (m1, m2) => m1 ++ m2) 
   }
 
 }
